requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
crispy-bootstrap5==2025.6
Django==5.2.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
sqlparse==0.5.3
